news-agent/requirements.txt
feedparser
beautifulsoup4
readability-lxml
requests
praw
google-generativeai
python-dotenv
schedule
python-dateutil
